Explicitly show kiosk mode is refreshing on inactivity timeout
Because the browser does the refresh in the background, and because it
takes a while, the interface is unresponsive for some seconds when the
kiosk mode is refreshing due to inactivity. This is unclear to the user
and the interface appears broken. It now shows a 'Refreshing' loader to
fix this incorrect perception.