In kiosk mode, add button to view user cart from swapped columns
When columns are swapped in kiosk mode, a list of users having items in
cart is shown along with a quantity number. There was no easy way to
list the items in a user cart. This change adds a button to that view,
to swap columns and view the given users cart contents.