Cancel successful BunqMeTab payments immediately
These payment requests weren't cancelled immediately before because it
caused a weird payment flow, where the request page showed an error
after paying. This only happened when the payment was made with a
monetary account on the same bunq account as the request was from. This
problem will probably never happen under normal circumstances. It is
therefore safer to immediately cancel the payment request instead.